+++ src/apps/dashboard-projects/Header/Sort/SortFunction.ts
@@ -1,0 +1,3 @@
+import type { ExtendedProjectData } from 'src/Types';
+
+export type SortFunction = (a: ExtendedProjectData, b: ExtendedProjectData) => number;
